fixed alembic problems, commented out FK dependencies to speed schema development
