+++ alembic.ini
@@ -1,6 +1,7 @@
 [alembic]
-script_location = alembic
-sqlalchemy.url = postgresql://user:password@host/dbname
+script_location = etl_merge/my_etl_project/alembic
+# The sqlalchemy.url is set dynamically in alembic/env.py
+# from the DATABASE_URL environment variable.
 
 [loggers]
 keys = root,sqlalchemy,alembic
