+++ etl_merge/my_etl_project/src/models/people_contacts.py
@@ -21,8 +21,8 @@ class AnalystContact(SQLModel, table=True):
     analyst_last_name: Optional[str] = Field(default=None)
     analyst_email: Optional[str] = Field(default=None)
     analyst_affiliation: Optional[int] = Field(default=None,
-                                               description="Reference to affiliations.affiliation_id",
-                                               foreign_key="affiliations.affiliation_id")
+                                               description="Reference to affiliations.affiliation_id")
+                                               # foreign_key="affiliations.affiliation_id")
 
 
 class Provider(SQLModel, table=True):
@@ -32,11 +32,11 @@ class Provider(SQLModel, table=True):
     provider_id: Optional[int] = Field(default=None, primary_key=True)
     provider_name: str = Field(..., unique=True, description="Not null")
     provider_affiliation: Optional[int] = Field(default=None,
-                                                description="Reference to affiliations.affiliation_id",
-                                                foreign_key="affiliations.affiliation_id")
+                                                description="Reference to affiliations.affiliation_id")
+                                                # foreign_key="affiliations.affiliation_id")
     provider_type_id: Optional[int] = Field(default=None,
-                                            description="Reference to provider_types.provider_type_id",
-                                            foreign_key="provider_types.provider_type_id")
+                                            description="Reference to provider_types.provider_type_id")
+                                            # foreign_key="provider_types.provider_type_id")
     anonymous: Optional[bool] = Field(default=None)
 
 
@@ -55,5 +55,5 @@ class Collector(SQLModel, table=True):
     collector_id: Optional[int] = Field(default=None, primary_key=True)
     collector_name: str = Field(..., unique=True, description="Not null")
     collector_affiliation: Optional[int] = Field(default=None,
-                                                 description="Reference to affiliations.affiliation_id",
-                                                 foreign_key="affiliations.affiliation_id")
+                                                 description="Reference to affiliations.affiliation_id")
+                                                 # foreign_key="affiliations.affiliation_id")
