+++ pixi.toml
@@ -20,6 +20,10 @@ python = "3.12.*"
 pre-commit = ">=4.2.0,<5"
 pytest = ">=8.4.2,<9"
 pytest-cov = ">=7.0.0,<8"
+sqlmodel = ">=0.0.19,<0.1"
+alembic = ">=1.13.2,<2"
+psycopg2 = ">=2.9.9,<3"
+python-dotenv = ">=1.0.1,<2"
 
 [tasks]
 pre-commit = "pre-commit run"
